+++ migrations/m140504_113157_update_tables.php
@@ -10,7 +10,6 @@
  */
 
 use dektrium\user\migrations\Migration;
-use yii\db\Schema;
 
 /**
  * @author Dmitry Erofeev <dmeroff@gmail.com>
@@ -20,8 +19,8 @@ class m140504_113157_update_tables extends Migration
     public function up()
     {
         // user table
-        $this->dropIndex('user_confirmation', '{{%user}}');
-        $this->dropIndex('user_recovery', '{{%user}}');
+        $this->dropIndex('{{%user_confirmation}}', '{{%user}}');
+        $this->dropIndex('{{%user_recovery}}', '{{%user}}');
         $this->dropColumn('{{%user}}', 'confirmation_token');
         $this->dropColumn('{{%user}}', 'confirmation_sent_at');
         $this->dropColumn('{{%user}}', 'recovery_token');
@@ -29,7 +28,7 @@ class m140504_113157_update_tables extends Migration
         $this->dropColumn('{{%user}}', 'logged_in_from');
         $this->dropColumn('{{%user}}', 'logged_in_at');
         $this->renameColumn('{{%user}}', 'registered_from', 'registration_ip');
-        $this->addColumn('{{%user}}', 'flags', Schema::TYPE_INTEGER . ' NOT NULL DEFAULT 0');
+        $this->addColumn('{{%user}}', 'flags', $this->integer()->notNull()->defaultValue(0));
 
         // account table
         $this->renameColumn('{{%account}}', 'properties', 'data');
@@ -41,15 +40,19 @@ class m140504_113157_update_tables extends Migration
         $this->renameColumn('{{%account}}', 'data', 'properties');
 
         // user table
+        if ($this->dbType == 'sqlsrv') {
+            // this is needed because we need to drop the default constraint first
+            $this->dropColumnConstraints('{{%user}}', 'flags');
+        }
         $this->dropColumn('{{%user}}', 'flags');
         $this->renameColumn('{{%user}}', 'registration_ip', 'registered_from');
-        $this->addColumn('{{%user}}', 'logged_in_at', Schema::TYPE_INTEGER);
-        $this->addColumn('{{%user}}', 'logged_in_from', Schema::TYPE_INTEGER);
-        $this->addColumn('{{%user}}', 'recovery_sent_at', Schema::TYPE_INTEGER);
-        $this->addColumn('{{%user}}', 'recovery_token', Schema::TYPE_STRING . '(32)');
-        $this->addColumn('{{%user}}', 'confirmation_sent_at', Schema::TYPE_INTEGER);
-        $this->addColumn('{{%user}}', 'confirmation_token', Schema::TYPE_STRING . '(32)');
-        $this->createIndex('user_confirmation', '{{%user}}', 'id, confirmation_token', true);
-        $this->createIndex('user_recovery', '{{%user}}', 'id, recovery_token', true);
+        $this->addColumn('{{%user}}', 'logged_in_at', $this->integer());
+        $this->addColumn('{{%user}}', 'logged_in_from', $this->integer());
+        $this->addColumn('{{%user}}', 'recovery_sent_at', $this->integer());
+        $this->addColumn('{{%user}}', 'recovery_token', $this->string(32));
+        $this->addColumn('{{%user}}', 'confirmation_sent_at', $this->integer());
+        $this->addColumn('{{%user}}', 'confirmation_token', $this->string(32));
+        $this->createIndex('{{%user_confirmation}}', '{{%user}}', 'id, confirmation_token', true);
+        $this->createIndex('{{%user_recovery}', '{{%user}}', 'id, recovery_token', true);
     }
 }