+++ migrations/m151218_234654_add_timezone_to_profile.php
@@ -9,19 +9,17 @@
  * file that was distributed with this source code.
  */
 
-use yii\db\Schema;
-use yii\db\Migration;
+use dektrium\user\migrations\Migration;
 
 class m151218_234654_add_timezone_to_profile extends Migration
 {
     public function up()
     {
-        $this->addColumn('{{%profile}}', 'timezone', Schema::TYPE_STRING . '(40)');
+        $this->addColumn('{{%profile}}', 'timezone', $this->string(40)->null());
     }
 
     public function down()
     {
         $this->dropcolumn('{{%profile}}', 'timezone');
     }
-
 }