+++ views/admin/_assignments.php
@@ -12,10 +12,9 @@
 use dektrium\rbac\widgets\Assignments;
 
 /**
- * @var yii\web\View                $this
- * @var dektrium\user\models\User   $user
+ * @var yii\web\View $this
+ * @var dektrium\user\models\User $user
  */
-
 ?>
 
 <?php $this->beginContent('@dektrium/user/views/admin/update.php', ['user' => $user]) ?>