+++ views/admin/index.php
@@ -9,35 +9,30 @@
  * file that was distributed with this source code.
  */
 
-use dektrium\user\models\UserSearch;
-use yii\data\ActiveDataProvider;
 use yii\grid\GridView;
 use yii\helpers\Html;
-use yii\web\View;
 use yii\widgets\Pjax;
 
 /**
- * @var View $this
- * @var ActiveDataProvider $dataProvider
- * @var UserSearch $searchModel
+ * @var \yii\web\View $this
+ * @var \yii\data\ActiveDataProvider $dataProvider
+ * @var \dektrium\user\models\UserSearch $searchModel
  */
 
 $this->title = Yii::t('user', 'Manage users');
 $this->params['breadcrumbs'][] = $this->title;
 ?>
 
-<?= $this->render('/_alert', [
-    'module' => Yii::$app->getModule('user'),
-]) ?>
+<?= $this->render('/_alert', ['module' => Yii::$app->getModule('user')]) ?>
 
 <?= $this->render('/admin/_menu') ?>
 
 <?php Pjax::begin() ?>
 
 <?= GridView::widget([
-    'dataProvider'  =>  $dataProvider,
-    'filterModel'   =>  $searchModel,
-    'layout'        =>  "{items}\n{pager}",
+    'dataProvider' => $dataProvider,
+    'filterModel' => $searchModel,
+    'layout' => "{items}\n{pager}",
     'columns' => [
         'username',
         'email:email',
@@ -99,7 +94,16 @@ $this->params['breadcrumbs'][] = $this->title;
         ],
         [
             'class' => 'yii\grid\ActionColumn',
-            'template' => '{update} {delete}',
+            'template' => '{switch} {update} {delete}',
+            'buttons' => [
+                'switch' => function ($url, $model) {
+                    if($model->id != Yii::$app->user->id)
+                        return Html::a('<span class="glyphicon glyphicon-user"></span>', ['/user/admin/switch', 'id' => $model->id], [
+                            'title' => Yii::t('user', 'Become this user'),
+                            'data-confirm' => Yii::t('user', 'Are you sure you want to switch to this user for the rest of this Session?'),
+                        ]);
+                }
+            ],
         ],
     ],
 ]); ?>