+++ views/message.php
@@ -10,14 +10,11 @@
  */
 
 /**
- * @var yii\web\View            $this
- * @var dektrium\user\Module    $module
+ * @var yii\web\View $this
+ * @var dektrium\user\Module $module
  */
 
 $this->title = $title;
-
 ?>
 
-<?= $this->render('/_alert', [
-    'module' => $module,
-]);
+<?= $this->render('/_alert', ['module' => $module]);