+++ views/registration/connect.php
@@ -13,10 +13,10 @@ use yii\helpers\Html;
 use yii\widgets\ActiveForm;
 
 /**
- * @var yii\web\View                    $this
- * @var yii\widgets\ActiveForm          $form
- * @var dektrium\user\models\User       $model
- * @var dektrium\user\models\Account    $account
+ * @var yii\web\View $this
+ * @var yii\widgets\ActiveForm $form
+ * @var dektrium\user\models\User $model
+ * @var dektrium\user\models\Account $account
  */
 
 $this->title = Yii::t('user', 'Sign in');